requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
langchain-google-genai
langchain-tavily
langchain_agentql
langchain-hugingface
langgraph
crawl4ai
sentence-transformers
pymupdf

playwright
lxml
beautifulsoup4

pydantic
pydantic[email]

ipykernel
python-dotenv

jinja2
fastapi "uvicorn[standard]"
python-multipart
sse-starlette